Created execute_option function
Broke the print_menu function into three parts a menu_option that validates the proper user inputs.
A print_menu function that only prints the menu and then calls menu_option to validate.
An execute_option function that calls the proper function on the cart class corresponding to the input.